assignment07c/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.txt
#
django==2.1
mysqlclient==1.3.13
pytz==2018.5              # via django
django-rest-swagger==2.2.0
djangorestframework==3.8.2